requirements.txt
streamlit
google-generativeai
python-dotenv
ipywidgets
pillow
pandas
transformers
torch
scikit-learn
chromadb
gradio
selenium
beautifulsoup4
lxml
html5lib
requests

